requirements.txt
Flask==3.1.3
PyJWT==2.11.0
bcrypt==5.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
crewai==0.102.0
crewai-tools
pydantic
